# docs/borga-api-spec-es.yaml
openapi: 3.0.0
info:
  description: Borga API
  version: 2.0.0
  title: Borga API Documentation
  contact:
    email: a47673@alunos.isel.pt
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /api/groups:
    get:
      tags:
        - groups
      summary: Get all the groups belonging the user
      description: Returns all the groups belonging to the user
      security:
        - bearerAuth: []
      responses:
        200:
          description: All the groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        401:
          description: Missing bearer token. The user is not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: No groups found for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
    post:
      tags:
        - groups
      summary: Adds a group
      description: Adds a group to the user
      security:
        - bearerAuth: []
      requestBody:
        description: Group to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
        required: true
      responses:
        201:
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        400:
          description: Invalid name and/or description provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: Missing bearer token. The user is not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /api/groups/{groupID}:
    get:
      tags:
        - groups
      summary: Get a group given by id
      description: Get a group given by id
      security:
        - bearerAuth: []
      parameters:
        - name: groupID
          in: path
          description: ID of group that to be fetched
          required: true
          schema:
            type: string
      responses:
        200:
          description: The found group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        400:
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: Missing bearer token. The user is not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: Group with that ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
    put:
      tags:
        - groups
      summary: Updates a group's name and description
      description: Updates a group in the system by changing its name and description
      security:
        - bearerAuth: []
      parameters:
        - name: groupID
          in: path
          description: ID of the group to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: New name and description for the group
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
        required: false
      responses:
        204:
          description: Group updated
        400:
          description: Invalid ID, name, and/or description provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: Missing bearer token. The user is not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: Group with that ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
    delete:
      tags:
        - groups
      summary: Delete a group by ID
      description: Delete a group by ID
      security:
        - bearerAuth: []
      parameters:
        - name: groupID
          in: path
          description: ID of the group to be deleted
          required: true
          schema:
            type: string
      responses:
        204:
          description: Group deleted
        400:
          description: Invalid group ID provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: Missing bearer token. The user is not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: Group with that ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    AppError:
      type: object
      description: Application error
      required:
        - description
      properties:
        description:
          type: string
          example: "Board Games Atlas API is not responding"
    SimpleGame:
      type: object
      required:
        - id
        - name
        - image_url
      properties:
        id:
          type: integer
        name:
          type: string
        image_url:
          type: string
...
    Group:
      type: object
      required:
        - id
        - name
        - description
        - games
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            $ref: '#/components/schemas/SimpleGame'
    NewGroup:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
...
servers:
  - url: https://isel-ipw-2122-1-g26d1.herokuapp.com/
    description: BORGA @ Heroku